# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO app Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "local server"
        }
    ],
    "paths": {
        "/task-categories": {
            "get": {
                "tags": [
                    "Task Category"
                ],
                "summary": "Get task categories list",
                "operationId": "53fffc44bfaf08594881d5fbe7a78c60",
                "responses": {
                    "200": {
                        "description": "Task categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task Category"
                ],
                "summary": "Create new task category",
                "operationId": "a9b125abaf498341c76ee590985e0f36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/StoreTaskCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task category created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/task-categories/{taskCategory}": {
            "get": {
                "tags": [
                    "Task Category"
                ],
                "summary": "Get task category",
                "operationId": "3ed5a7aa0cf1a555744234c443287b35",
                "responses": {
                    "200": {
                        "description": "Task category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskCategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task Category"
                ],
                "summary": "Update task category",
                "operationId": "9e98ded6b5d3ecb9c7d0d4c1db36cb8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/UpdateTaskCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task category updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task Category"
                ],
                "summary": "Delete task category",
                "operationId": "cf8b2f34952f0399f7a9eb0413df7d2d",
                "responses": {
                    "200": {
                        "description": "Task category deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get tasks list",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create new task",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get task",
                "operationId": "2d3bd1c8c8bbb942e64655b95d2ccd31",
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update task",
                "operationId": "21fde229ffe8d43f5907fdc299a6e0be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "operationId": "2e76d38ee629d5be516c498ed02ff818",
                "responses": {
                    "200": {
                        "description": "Task deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TaskCategoryResource": {
                "title": "TaskCategoryResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "user": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ]
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskCategoryResource"
                }
            },
            "TaskResource": {
                "title": "TaskResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "durationMinutes": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "userId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "taskCategoryId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "user": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ]
                    },
                    "taskCategory": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/TaskCategoryResource"
                            }
                        ]
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskResource"
                }
            },
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}